+++ src/docs/pages/breadcrumbs-demo/breadcrumbs-demo.ts
@@ -10,5 +10,5 @@ export class BreadcrumbsDemo {
     isChanged: boolean = false;
 
     multiline: boolean = true;
-    multilineExpanded: boolean = true;
+    multilineExpanded: boolean = false;
 }